[tests] Upgrade to maven plugin testing harness 4.0.0-alpha-2 but still include their code
There code didn't have direct access for what we need still so just pull updates to their actual release this time.  Original commit was from June 18, 2022.

For 4.0.0 alpha 1 they really only added generics as seen here and made it compatible with maven 4.

For 4.0.0 alpha 2 they only made it more compatible with maven 3 and made their code junit 5 basically like we did where it still requires junit 4 for plexus.

Therefore this was not as good as hoped to remove the code but non the less we got rid of a lot of libraries.